pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autotradex"
version = "0.2.0"
description = "Self-Evolving Crypto Trading Agents with CoinGecko MCP Integration"
readme = "README.md"
authors = [
    {name = "AutoTradeX Team", email = "info@autotradex.io"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "Topic :: Office/Business :: Financial :: Investment",
]
requires-python = ">=3.10"
dependencies = [
    "groq>=0.4.1",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "qdrant-client>=1.7.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
    "requests>=2.31.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "python-json-logger>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.1.0",
    "mypy>=1.5.0",
    "isort>=5.12.0",
]
rl = [
    "stable-baselines3>=2.0.0",
    "gymnasium>=0.28.1",
]
langgraph = [
    "langchain-core>=0.2.0",
    "langchain-groq>=0.1.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/autotradex"
"Bug Tracker" = "https://github.com/yourusername/autotradex/issues"
"Documentation" = "https://autotradex.readthedocs.io"

[project.scripts]
autotradex = "autotradex.cli:main"

[tool.setuptools]
package-dir = {"" = "."}
packages = ["autotradex"]

[tool.black]
line-length = 88
target-version = ["py310", "py311"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*"
addopts = "--cov=autotradex"
